Update compound.java
Create a program that will calculate the compound interest. You must use 2 or more classes or 2 or more methods to solve the problem. 
From the user get the invested amount, rate of interest, and the number of years. Based on that information, make a program to find the amount that you will receive at the end of the year.